Refactor: extract shared behaviour for scraping IO from a class method
