+++ pyiron_workflow/io.py
@@ -7,9 +7,10 @@ connections on the same footing.
 
 from __future__ import annotations
 
-import warnings
 from abc import ABC, abstractmethod
-from typing import Any
+import inspect
+from typing import Any, get_args, get_type_hints
+import warnings
 
 from pyiron_workflow.channels import (
     Channel,
@@ -24,6 +25,7 @@ from pyiron_workflow.channels import (
 )
 from pyiron_workflow.has_interface_mixins import HasChannel, HasLabel, HasRun, UsesState
 from pyiron_workflow.has_to_dict import HasToDict
+from pyiron_workflow.output_parser import ParseOutput
 from pyiron_workflow.snippets.logger import logger
 from pyiron_workflow.snippets.dotdict import DotDict
 
@@ -575,3 +577,190 @@ class ConnectionCopyError(ValueError):
 
 class ValueCopyError(ValueError):
     """Raised when trying to copy IO, but values cannot be copied"""
+
+
+class ScrapesIO(ABC):
+    """
+    A mixin class for scraping IO channel information from a specific class method.
+
+    Requires that the (static and class) method :meth:`_io_defining_function` be
+    specified in child classes.
+
+    Class methods:
+        preview_input_channels (dict[str, tuple[Any, Any]]): Input channel names paired
+            with their type hint (if any, may be `None`) and default value (if any,
+            may be `pyiron_workflow.channels.NOT_DATA`).
+        preview_output_channels (dict[str, Any]): Output channel names paired with
+            their type hint (if any, may be `None`). Channel names are scr
+
+    Warning:
+        There are a number of class features which, for computational efficiency, get
+        calculated at first call and any subsequent calls return that initial value
+        (including on other instances, since these are class properties); these
+        depend on the :meth:`_io_defining_function` and its signature, which should
+        thus be left static from the time of class definition onwards.
+    """
+
+    @classmethod
+    @abstractmethod
+    def _io_defining_function(cls) -> callable:
+        """Must return a static class method."""
+
+    @classmethod
+    @abstractmethod
+    def _io_defining_function_uses_self(cls) -> bool:
+        """Whether the signature of the IO defining function starts with self."""
+
+    _provided_output_labels: tuple[str] | None = None
+    __type_hints = None
+    __input_args = None
+    __init_keywords = None
+
+    @classmethod
+    def preview_input_channels(cls) -> dict[str, tuple[Any, Any]]:
+        """
+        Gives a class-level peek at the expected input channels.
+
+        Returns:
+            dict[str, tuple[Any, Any]]: The channel name and a tuple of its
+                corresponding type hint and default value.
+        """
+        type_hints = cls._type_hints()
+        scraped: dict[str, tuple[Any, Any]] = {}
+        for i, (label, value) in enumerate(cls._input_args().items()):
+            if cls._io_defining_function_uses_self() and i == 0:
+                continue  # Skip the macro argument itself, it's like `self` here
+            elif label in cls._init_keywords():
+                # We allow users to parse arbitrary kwargs as channel initialization
+                # So don't let them choose bad channel names
+                raise ValueError(
+                    f"The Input channel name {label} is not valid. Please choose a "
+                    f"name _not_ among {cls._init_keywords()}"
+                )
+
+            try:
+                type_hint = type_hints[label]
+            except KeyError:
+                type_hint = None
+
+            default = (
+                value.default
+                if value.default is not inspect.Parameter.empty
+                else NOT_DATA
+            )
+
+            scraped[label] = (type_hint, default)
+        return scraped
+
+    @classmethod
+    def preview_output_channels(cls) -> dict[str, Any]:
+        """
+        Gives a class-level peek at the expected output channels.
+
+        Returns:
+            dict[str, tuple[Any, Any]]: The channel name and its corresponding type
+                hint.
+        """
+        labels = cls._get_output_labels()
+        labels = [] if labels is None else labels
+        try:
+            type_hints = cls._type_hints()["return"]
+            if len(labels) > 1:
+                type_hints = get_args(type_hints)
+                if not isinstance(type_hints, tuple):
+                    raise TypeError(
+                        f"With multiple return labels expected to get a tuple of type "
+                        f"hints, but got type {type(type_hints)}"
+                    )
+                if len(type_hints) != len(labels):
+                    raise ValueError(
+                        f"Expected type hints and return labels to have matching "
+                        f"lengths, but got {len(type_hints)} hints and "
+                        f"{len(labels)} labels: {type_hints}, {labels}"
+                    )
+            else:
+                # If there's only one hint, wrap it in a tuple, so we can zip it with
+                # *return_labels and iterate over both at once
+                type_hints = (type_hints,)
+        except KeyError:  # If there are no return hints
+            type_hints = [None] * len(labels)
+            # Note that this nicely differs from `NoneType`, which is the hint when
+            # `None` is actually the hint!
+        return {label: hint for label, hint in zip(labels, type_hints)}
+
+    @classmethod
+    def _get_output_labels(cls):
+        """
+        Return output labels provided for the class, scraping them from the io-defining
+        function if they are not already available.
+        """
+        if cls._provided_output_labels is None:
+            cls._provided_output_labels = cls._scrape_output_labels()
+        return cls._provided_output_labels
+
+    @classmethod
+    def _type_hints(cls) -> dict:
+        """
+        The result of :func:`typing.get_type_hints` on the io-defining function
+        """
+        if cls.__type_hints is None:
+            cls.__type_hints = get_type_hints(cls._io_defining_function())
+        return cls.__type_hints
+
+    @classmethod
+    def _input_args(cls):
+        if cls.__input_args is None:
+            cls.__input_args = inspect.signature(cls._io_defining_function()).parameters
+        return cls.__input_args
+
+    @classmethod
+    def _init_keywords(cls):
+        if cls.__init_keywords is None:
+            cls.__init_keywords = list(
+                inspect.signature(cls.__init__).parameters.keys()
+            )
+        return cls.__init_keywords
+
+    @classmethod
+    def _scrape_output_labels(cls):
+        """
+        Inspect :meth:`node_function` to scrape out strings representing the
+        returned values.
+
+         _Only_ works for functions with a single `return` expression in their body.
+
+        It will return expressions and function calls just fine, thus good practice is
+        to create well-named variables and return those so that the output labels stay
+        dot-accessible.
+        """
+        return ParseOutput(cls._io_defining_function()).output
+
+    @classmethod
+    def _macro_validate_output_labels(cls) -> tuple[str]:
+        """
+        Ensure that output_labels, if provided, are commensurate with graph creator
+        return values, if provided, and return them as a tuple.
+        """
+        graph_creator_returns = ParseOutput(cls._io_defining_function()).output
+        output_labels = cls._get_output_labels()
+        if output_labels is not None and len(set(output_labels)) != len(output_labels):
+            raise ValueError(
+                f"{cls.__name__} must not have degenerate output labels: "
+                f"{output_labels}"
+            )
+        if graph_creator_returns is not None or output_labels is not None:
+            error_suffix = (
+                f"but {cls.__name__} macro class got return values: "
+                f"{graph_creator_returns} and labels: {output_labels}."
+            )
+            try:
+                if len(output_labels) != len(graph_creator_returns):
+                    raise ValueError(
+                        "The number of return values in the graph creator must exactly "
+                        "match the number of output labels provided, " + error_suffix
+                    )
+            except TypeError:
+                raise TypeError(
+                    f"Output labels and graph creator return values must either both "
+                    f"or neither be present, " + error_suffix
+                )
