+++ pyiron_workflow/macro.py
@@ -6,22 +6,20 @@ interface and are not intended to be internally modified after instantiation.
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-import inspect
 import re
-from typing import Any, get_args, get_type_hints, Literal, Optional, TYPE_CHECKING
+from typing import Literal, Optional, TYPE_CHECKING
 import warnings
 
-from pyiron_workflow.channels import InputData, OutputData, NOT_DATA
+from pyiron_workflow.channels import InputData, OutputData
 from pyiron_workflow.composite import Composite
 from pyiron_workflow.has_interface_mixins import HasChannel
-from pyiron_workflow.io import Outputs, Inputs
-from pyiron_workflow.output_parser import ParseOutput
+from pyiron_workflow.io import Outputs, Inputs, ScrapesIO
 
 if TYPE_CHECKING:
     from pyiron_workflow.channels import Channel
 
 
-class Macro(Composite, ABC):
+class Macro(Composite, ScrapesIO, ABC):
     """
     A macro is a composite node that holds a graph with a fixed interface, like a
     pre-populated workflow that is the same every time you instantiate it.
@@ -245,8 +243,6 @@ class Macro(Composite, ABC):
 
     """
 
-    _provided_output_labels: tuple[str] | None = None
-
     def __init__(
         self,
         label: Optional[str] = None,
@@ -307,105 +303,23 @@ class Macro(Composite, ABC):
         """Build the graph the node will run."""
 
     @classmethod
-    def _validate_output_labels(cls) -> tuple[str]:
-        """
-        Ensure that output_labels, if provided, are commensurate with graph creator
-        return values, if provided, and return them as a tuple.
-        """
-        graph_creator_returns = ParseOutput(cls.graph_creator).output
-        output_labels = cls._get_output_labels()
-        if output_labels is not None and len(set(output_labels)) != len(output_labels):
-            raise ValueError(
-                f"{cls.__name__} must not have degenerate output labels: "
-                f"{output_labels}"
-            )
-        if graph_creator_returns is not None or output_labels is not None:
-            error_suffix = (
-                f"but {cls.__name__} macro class got return values: "
-                f"{graph_creator_returns} and labels: {output_labels}."
-            )
-            try:
-                if len(output_labels) != len(graph_creator_returns):
-                    raise ValueError(
-                        "The number of return values in the graph creator must exactly "
-                        "match the number of output labels provided, " + error_suffix
-                    )
-            except TypeError:
-                raise TypeError(
-                    f"Output labels and graph creator return values must either both "
-                    f"or neither be present, " + error_suffix
-                )
+    def _io_defining_function(cls) -> callable:
+        return cls.graph_creator
 
     @classmethod
-    def _type_hints(cls):
-        """The result of :func:`typing.get_type_hints` on the :meth:`graph_creator`."""
-        return get_type_hints(cls.graph_creator)
-
-    @classmethod
-    def preview_output_channels(cls) -> dict[str, Any]:
-        """
-        Gives a class-level peek at the expected output channels.
-
-        Returns:
-            dict[str, tuple[Any, Any]]: The channel name and its corresponding type
-                hint.
-        """
-        labels = cls._get_output_labels()
-        try:
-            type_hints = cls._type_hints()["return"]
-            if len(labels) > 1:
-                type_hints = get_args(type_hints)
-                if not isinstance(type_hints, tuple):
-                    raise TypeError(
-                        f"With multiple return labels expected to get a tuple of type "
-                        f"hints, but got type {type(type_hints)}"
-                    )
-                if len(type_hints) != len(labels):
-                    raise ValueError(
-                        f"Expected type hints and return labels to have matching "
-                        f"lengths, but got {len(type_hints)} hints and "
-                        f"{len(labels)} labels: {type_hints}, {labels}"
-                    )
-            else:
-                # If there's only one hint, wrap it in a tuple, so we can zip it with
-                # *return_labels and iterate over both at once
-                type_hints = (type_hints,)
-        except KeyError:  # If there are no return hints
-            type_hints = [None] * len(labels)
-            # Note that this nicely differs from `NoneType`, which is the hint when
-            # `None` is actually the hint!
-        return {label: hint for label, hint in zip(labels, type_hints)}
-
-    @classmethod
-    def _get_output_labels(cls):
-        """
-        Return output labels provided on the class if not None.
-        """
-        if cls._provided_output_labels is None:
-            cls._scrape_output_labels()
-        return cls._provided_output_labels
+    def _io_defining_function_uses_self(cls) -> callable:
+        return True
 
     @classmethod
     def _scrape_output_labels(cls):
-        """
-        Inspect :meth:`node_function` to scrape out strings representing the
-        returned values.
+        scraped_labels = super(Macro, cls)._scrape_output_labels()
 
-         _Only_ works for functions with a single `return` expression in their body.
-
-        It will return expressions and function calls just fine, thus good practice is
-        to create well-named variables and return those so that the output labels stay
-        dot-accessible.
-        """
-        parsed_outputs = ParseOutput(cls.graph_creator).output
-        if parsed_outputs is None:
-            cls._provided_output_labels = None
-        else:
+        if scraped_labels is not None:
+            # Strip off the first argument, e.g. self.foo just becomes foo
             self_argument = list(cls._input_args().keys())[0]
             cleaned_labels = [
-                # Strip off the first argument, e.g. self.foo just becomes foo
-                re.sub(r"^" + re.escape(f"{self_argument}."), "", p)
-                for p in parsed_outputs
+                re.sub(r"^" + re.escape(f"{self_argument}."), "", label)
+                for label in scraped_labels
             ]
             if any("." in label for label in cleaned_labels):
                 raise ValueError(
@@ -413,51 +327,9 @@ class Macro(Composite, ABC):
                     f"one of {cleaned_labels} still contains a '.' -- please provide "
                     f"explicit labels"
                 )
-            cls._provided_output_labels = cleaned_labels
-
-    @classmethod
-    def preview_input_channels(cls) -> dict[str, tuple[Any, Any]]:
-        """
-        Gives a class-level peek at the expected input channels.
-
-        Returns:
-            dict[str, tuple[Any, Any]]: The channel name and a tuple of its
-                corresponding type hint and default value.
-        """
-        type_hints = cls._type_hints()
-        scraped: dict[str, tuple[Any, Any]] = {}
-        for i, (label, value) in enumerate(cls._input_args().items()):
-            if i == 0:
-                continue  # Skip the macro argument itself, it's like `self` here
-            elif label in cls._init_keywords():
-                # We allow users to parse arbitrary kwargs as channel initialization
-                # So don't let them choose bad channel names
-                raise ValueError(
-                    f"The Input channel name {label} is not valid. Please choose a "
-                    f"name _not_ among {cls._init_keywords()}"
-                )
-
-            try:
-                type_hint = type_hints[label]
-            except KeyError:
-                type_hint = None
-
-            default = (
-                value.default
-                if value.default is not inspect.Parameter.empty
-                else NOT_DATA
-            )
-
-            scraped[label] = (type_hint, default)
-        return scraped
-
-    @classmethod
-    def _input_args(cls):
-        return inspect.signature(cls.graph_creator).parameters
-
-    @classmethod
-    def _init_keywords(cls):
-        return list(inspect.signature(cls.__init__).parameters.keys())
+            return cleaned_labels
+        else:
+            return scraped_labels
 
     def _prepopulate_ui_nodes_from_graph_creator_signature(
         self, storage_backend: Literal["h5io", "tinybase"]
@@ -732,7 +604,7 @@ def as_macro_node(*output_labels):
             },
         )
         try:
-            node_class._validate_output_labels()
+            node_class._macro_validate_output_labels()
         except OSError:
             warnings.warn(
                 f"Could not find the source code to validate {node_class.__name__} "
